+++ bench/src/run_bench_avro.rs
@@ -10,10 +10,11 @@ use serde::{Deserialize, Serialize};
 use std::io::prelude::*;
 use std::time::SystemTime;
 
-
 #[repr(i32)]
 enum Fruit {
-    Apples, Pears, Bananas
+    Apples,
+    Pears,
+    Bananas,
 }
 
 pub struct AvroBench();
@@ -128,7 +129,6 @@ impl AvroBench {
     }
 
     pub fn update_bench(base: u128) -> String {
-
         let schema = Self::get_schema();
 
         let buffer = Self::encode_single(&schema);
@@ -136,41 +136,41 @@ impl AvroBench {
         let start = SystemTime::now();
 
         for _x in 0..LOOPS {
-
             let reader = Reader::new(&buffer[..]).unwrap();
             let mut foobar_c: Vec<(String, Value)> = Vec::new();
-
-            
 
             for val in reader {
                 if let Value::Record(data) = val.unwrap() {
                     data.iter().for_each(|(key, data)| {
-						match key.as_str() {
+                        match key.as_str() {
                             "list" => {
-
                                 let mut vector: Vec<Value> = Vec::new();
 
                                 if let Value::Array(list) = data {
                                     list.iter().enumerate().for_each(|(i, foo_bar)| {
-                                        
                                         if i == 0 {
                                             if let Value::Record(foo_bar) = foo_bar {
-                                                let mut new_foobar: Vec<(String, Value)> = Vec::new();
+                                                let mut new_foobar: Vec<(String, Value)> =
+                                                    Vec::new();
 
                                                 foo_bar.iter().for_each(|(key, value)| {
                                                     match key.as_str() {
                                                         "name" => {
-                                                            new_foobar.push((String::from("name"), Value::String(String::from("bob"))));
-                                                        },
+                                                            new_foobar.push((
+                                                                String::from("name"),
+                                                                Value::String(String::from("bob")),
+                                                            ));
+                                                        }
                                                         _ => {
-                                                            new_foobar.push((key.clone(), value.clone()));
+                                                            new_foobar
+                                                                .push((key.clone(), value.clone()));
                                                         }
                                                     }
                                                 });
                                                 vector.push(Value::Record(new_foobar));
                                             } else {
                                                 panic!()
-                                            }                                            
+                                            }
                                         } else {
                                             vector.push(foo_bar.clone());
                                         }
@@ -180,11 +180,11 @@ impl AvroBench {
                                 }
 
                                 foobar_c.push((String::from("list"), Value::Array(vector)));
-                            },
+                            }
                             _ => {
                                 foobar_c.push((key.clone(), data.clone()));
                             }
-						};
+                        };
                     });
                 }
             }
@@ -195,14 +195,18 @@ impl AvroBench {
             assert_eq!(finished.len(), 692);
         }
 
-        let time = SystemTime::now().duration_since(start).expect("Time went backwards");
-        println!("Avro:        {:>9.0} ops/ms {:.2}", LOOPS as f64 / time.as_millis() as f64, (base as f64 / time.as_micros() as f64));
+        let time = SystemTime::now()
+            .duration_since(start)
+            .expect("Time went backwards");
+        println!(
+            "Avro:        {:>9.0} ops/ms {:.2}",
+            LOOPS as f64 / time.as_millis() as f64,
+            (base as f64 / time.as_micros() as f64)
+        );
         format!("{:>6.0}", LOOPS as f64 / time.as_millis() as f64)
     }
 
     pub fn decode_one_bench(base: u128) -> String {
-        
-
         let schema = Self::get_schema();
 
         let start = SystemTime::now();
@@ -246,36 +250,35 @@ impl AvroBench {
         let start = SystemTime::now();
 
         for _x in 0..LOOPS {
-
             let mut loops = 0;
 
             let reader = Reader::new(&buffer[..]).unwrap();
             for val in reader {
                 if let Value::Record(data) = val.unwrap() {
                     data.iter().for_each(|(key, data)| {
-						match key.as_str() {
-							"location" => {
+                        match key.as_str() {
+                            "location" => {
                                 if let Value::String(x) = data {
                                     assert_eq!("http://arstechnica.com", x);
                                 } else {
                                     panic!()
                                 }
-							},
-							"initialized" => {
+                            }
+                            "initialized" => {
                                 if let Value::Boolean(x) = data {
                                     assert_eq!(true, *x);
                                 } else {
                                     panic!()
                                 }
-							},
-							"fruit" => {
+                            }
+                            "fruit" => {
                                 if let Value::Int(x) = data {
                                     assert_eq!(Fruit::Apples as i32, *x);
                                 } else {
                                     panic!()
                                 }
-							},
-							"list" => {
+                            }
+                            "list" => {
                                 if let Value::Array(list) = data {
                                     list.iter().for_each(|foo_bar| {
                                         if let Value::Record(foo_bar) = foo_bar {
@@ -289,45 +292,72 @@ impl AvroBench {
                                                         } else {
                                                             panic!()
                                                         }
-                                                    },
+                                                    }
                                                     "rating" => {
                                                         if let Value::Float(x) = value {
-                                                            assert_eq!(3.1415432432445543543 + (loops as f32), *x);
+                                                            assert_eq!(
+                                                                3.1415432432445543543
+                                                                    + (loops as f32),
+                                                                *x
+                                                            );
                                                         } else {
                                                             panic!()
                                                         }
-                                                    },
+                                                    }
                                                     "postfix" => {
                                                         if let Value::String(x) = value {
                                                             assert_eq!("!", x);
                                                         } else {
                                                             panic!()
                                                         }
-                                                    },
+                                                    }
                                                     "sibling" => {
                                                         if let Value::Record(sibling) = value {
                                                             let mut foo_key_count = 0;
-                                                            sibling.iter().for_each(|(skey, svalue)| {
-                                                                foo_key_count += 1;
-                                                                match skey.as_str() {
-                                                                    "time" => {
-                                                                        assert_eq!(Value::Int(123456 + (loops as i32)), *svalue);
-                                                                    },
-                                                                    "ratio" => {
-                                                                        assert_eq!(Value::Float(3.14159 + (loops as f32)), *svalue);
-                                                                    },
-                                                                    "size" => {
-                                                                        assert_eq!(Value::Int(10000 + (loops as i32)), *svalue);
-                                                                    },
-                                                                    _ => panic!()
-                                                                }
-                                                            });
+                                                            sibling.iter().for_each(
+                                                                |(skey, svalue)| {
+                                                                    foo_key_count += 1;
+                                                                    match skey.as_str() {
+                                                                        "time" => {
+                                                                            assert_eq!(
+                                                                                Value::Int(
+                                                                                    123456
+                                                                                        + (loops
+                                                                                            as i32)
+                                                                                ),
+                                                                                *svalue
+                                                                            );
+                                                                        }
+                                                                        "ratio" => {
+                                                                            assert_eq!(
+                                                                                Value::Float(
+                                                                                    3.14159
+                                                                                        + (loops
+                                                                                            as f32)
+                                                                                ),
+                                                                                *svalue
+                                                                            );
+                                                                        }
+                                                                        "size" => {
+                                                                            assert_eq!(
+                                                                                Value::Int(
+                                                                                    10000
+                                                                                        + (loops
+                                                                                            as i32)
+                                                                                ),
+                                                                                *svalue
+                                                                            );
+                                                                        }
+                                                                        _ => panic!(),
+                                                                    }
+                                                                },
+                                                            );
                                                             assert_eq!(foo_key_count, 3);
                                                         } else {
                                                             panic!()
                                                         }
-                                                    },
-                                                    _ => panic!()
+                                                    }
+                                                    _ => panic!(),
                                                 }
                                             });
                                             assert_eq!(key_count, 4);
@@ -340,9 +370,9 @@ impl AvroBench {
                                 } else {
                                     panic!()
                                 }
-							},
-							_ => panic!()
-						};
+                            }
+                            _ => panic!(),
+                        };
                     });
                 }
             }
@@ -350,8 +380,14 @@ impl AvroBench {
             assert!(loops == 3);
         }
 
-        let time = SystemTime::now().duration_since(start).expect("Time went backwards");
-        println!("Avro:        {:>9.0} ops/ms {:.2}", LOOPS as f64 / time.as_millis() as f64, (base as f64 / time.as_micros() as f64));
+        let time = SystemTime::now()
+            .duration_since(start)
+            .expect("Time went backwards");
+        println!(
+            "Avro:        {:>9.0} ops/ms {:.2}",
+            LOOPS as f64 / time.as_millis() as f64,
+            (base as f64 / time.as_micros() as f64)
+        );
         format!("{:>6.0}", LOOPS as f64 / time.as_millis() as f64)
     }
 }