+++ no_proto_rs/src/format.rs
@@ -864,3 +864,7 @@
 //!
 //! # Ok::<(), NP_Error>(()) 
 //! ```
+//! 
+//! ### Portal (Collection/Scalar)
+//! 
+//!