+++ no_proto_rs/src/schema.rs
@@ -33,10 +33,16 @@
 //!     values?: NP_Schema[]
 //! 
 //!     // used by table types
-//!     columns?: [string, NP_Schema][]
+//!     columns?: [string, NP_Schema][];
 //! 
 //!     // used by option/enum types
 //!     choices?: string[];
+//!     
+//!     // used by unions
+//!     types?: [string, NP_Schema][];
+//!     
+//!     // used by portals
+//!     to?: string
 //! 
 //!     // default value for this item
 //!     default?: any;
@@ -559,7 +565,7 @@
 //! [Go to NP_Factory docs](../struct.NP_Factory.html)
 //! 
 
-use crate::pointer::NP_Cursor;
+use crate::pointer::{NP_Cursor, union::NP_Union};
 use alloc::string::String;
 use core::{fmt::Debug};
 use crate::{buffer::DEFAULT_ROOT_PTR_ADDR, json_flex::NP_JSON, memory::NP_Memory_Writable, pointer::{portal::{NP_Portal}, ulid::NP_ULID, uuid::NP_UUID}};
@@ -580,37 +586,38 @@ use alloc::boxed::Box;
 #[repr(u8)]
 #[allow(missing_docs)]
 pub enum NP_TypeKeys {
-    None = 0,
-    Any = 1,
-    UTF8String = 2,
-    Bytes = 3,
-    Int8 = 4,
-    Int16 = 5,
-    Int32 = 6,
-    Int64 = 7,
-    Uint8 = 8,
-    Uint16 = 9,
-    Uint32 = 10,
-    Uint64 = 11,
-    Float = 12,
-    Double = 13,
-    Decimal = 14,
-    Boolean = 15,
-    Geo = 16,
-    Uuid = 17,
-    Ulid = 18,
-    Date = 19,
-    Enum = 20,
-    Table = 21,
-    Map = 22, 
-    List = 23,
-    Tuple = 24,
-    Portal = 25
+    None       =  0,
+    Any        =  1,
+    UTF8String =  2,
+    Bytes      =  3,
+    Int8       =  4,
+    Int16      =  5,
+    Int32      =  6,
+    Int64      =  7,
+    Uint8      =  8,
+    Uint16     =  9,
+    Uint32     = 10,
+    Uint64     = 11,
+    Float      = 12,
+    Double     = 13,
+    Decimal    = 14,
+    Boolean    = 15,
+    Geo        = 16,
+    Uuid       = 17,
+    Ulid       = 18,
+    Date       = 19,
+    Enum       = 20,
+    Table      = 21,
+    Map        = 22, 
+    List       = 23,
+    Tuple      = 24,
+    Portal     = 25,
+    Union      = 26
 }
 
 impl From<u8> for NP_TypeKeys {
     fn from(value: u8) -> Self {
-        if value > 25 { return NP_TypeKeys::None; }
+        if value > 26 { return NP_TypeKeys::None; }
         unsafe { core::mem::transmute(value) }
     }
 }
@@ -619,31 +626,31 @@ impl NP_TypeKeys {
     /// Convert this NP_TypeKey into a specific type index
     pub fn into_type_idx<'idx>(&self) -> (&'idx str, NP_TypeKeys) {
         match self {
-            NP_TypeKeys::None =>       {    ("none", NP_TypeKeys::None) }
-            NP_TypeKeys::Any =>        {    NP_Any::type_idx() }
+            NP_TypeKeys::None       => {    ("none", NP_TypeKeys::None) }
+            NP_TypeKeys::Any        => {    NP_Any::type_idx() }
             NP_TypeKeys::UTF8String => {    String::type_idx() }
-            NP_TypeKeys::Bytes =>      {  NP_Bytes::type_idx() }
-            NP_TypeKeys::Int8 =>       {        i8::type_idx() }
-            NP_TypeKeys::Int16 =>      {       i16::type_idx() }
-            NP_TypeKeys::Int32 =>      {       i32::type_idx() }
-            NP_TypeKeys::Int64 =>      {       i64::type_idx() }
-            NP_TypeKeys::Uint8 =>      {        u8::type_idx() }
-            NP_TypeKeys::Uint16 =>     {       u16::type_idx() }
-            NP_TypeKeys::Uint32 =>     {       u32::type_idx() }
-            NP_TypeKeys::Uint64 =>     {       u64::type_idx() }
-            NP_TypeKeys::Float =>      {       f32::type_idx() }
-            NP_TypeKeys::Double =>     {       f64::type_idx() }
-            NP_TypeKeys::Decimal =>    {    NP_Dec::type_idx() }
-            NP_TypeKeys::Boolean =>    {      bool::type_idx() }
-            NP_TypeKeys::Geo =>        {    NP_Geo::type_idx() }
-            NP_TypeKeys::Uuid =>       {   NP_UUID::type_idx() }
-            NP_TypeKeys::Ulid =>       {   NP_ULID::type_idx() }
-            NP_TypeKeys::Date =>       {   NP_Date::type_idx() }
-            NP_TypeKeys::Enum =>       {   NP_Enum::type_idx() }
-            NP_TypeKeys::Table =>      {  NP_Table::type_idx() }
-            NP_TypeKeys::Map =>        {    NP_Map::type_idx() }
-            NP_TypeKeys::List =>       {   NP_List::type_idx() }
-            NP_TypeKeys::Tuple =>      {  NP_Tuple::type_idx() },
+            NP_TypeKeys::Bytes      => {  NP_Bytes::type_idx() }
+            NP_TypeKeys::Int8       => {        i8::type_idx() }
+            NP_TypeKeys::Int16      => {       i16::type_idx() }
+            NP_TypeKeys::Int32      => {       i32::type_idx() }
+            NP_TypeKeys::Int64      => {       i64::type_idx() }
+            NP_TypeKeys::Uint8      => {        u8::type_idx() }
+            NP_TypeKeys::Uint16     => {       u16::type_idx() }
+            NP_TypeKeys::Uint32     => {       u32::type_idx() }
+            NP_TypeKeys::Uint64     => {       u64::type_idx() }
+            NP_TypeKeys::Float      => {       f32::type_idx() }
+            NP_TypeKeys::Double     => {       f64::type_idx() }
+            NP_TypeKeys::Decimal    => {    NP_Dec::type_idx() }
+            NP_TypeKeys::Boolean    => {      bool::type_idx() }
+            NP_TypeKeys::Geo        => {    NP_Geo::type_idx() }
+            NP_TypeKeys::Uuid       => {   NP_UUID::type_idx() }
+            NP_TypeKeys::Ulid       => {   NP_ULID::type_idx() }
+            NP_TypeKeys::Date       => {   NP_Date::type_idx() }
+            NP_TypeKeys::Enum       => {   NP_Enum::type_idx() }
+            NP_TypeKeys::Table      => {  NP_Table::type_idx() }
+            NP_TypeKeys::Map        => {    NP_Map::type_idx() }
+            NP_TypeKeys::List       => {   NP_List::type_idx() }
+            NP_TypeKeys::Tuple      => {  NP_Tuple::type_idx() },
             _ => ("", NP_TypeKeys::None)
         }
     }
@@ -698,7 +705,8 @@ pub enum NP_Parsed_Schema {
     Map        { sortable: bool, i:NP_TypeKeys, value: NP_Schema_Addr}, 
     List       { sortable: bool, i:NP_TypeKeys, of: NP_Schema_Addr },
     Tuple      { sortable: bool, i:NP_TypeKeys, values: Vec<NP_Schema_Addr>},
-    Portal     { sortable: bool, i:NP_TypeKeys, path: String, schema: usize, parent_schema: usize }
+    Portal     { sortable: bool, i:NP_TypeKeys, path: String, schema: usize, parent_schema: usize },
+    Union      { sortable: bool, i:NP_TypeKeys, types: Vec<(u8, String, NP_Schema_Addr)>, default: usize },
 }
 
 impl NP_Parsed_Schema {
@@ -731,7 +739,8 @@ impl NP_Parsed_Schema {
             NP_Parsed_Schema::Map        { i, .. }     => { i }
             NP_Parsed_Schema::List       { i, .. }     => { i }
             NP_Parsed_Schema::Tuple      { i, .. }     => { i }
-            NP_Parsed_Schema::Portal  { i, .. }     => { i }
+            NP_Parsed_Schema::Portal     { i, .. }     => { i }
+            NP_Parsed_Schema::Union      { i, .. }     => { i }
         }
     }
 
@@ -739,7 +748,7 @@ impl NP_Parsed_Schema {
     pub fn get_type_data(&self) -> (&str, NP_TypeKeys) {
         match self {
             NP_Parsed_Schema::None =>      ("", NP_TypeKeys::None),
-            NP_Parsed_Schema::Portal  { i, .. }     => { i.into_type_idx() }
+            NP_Parsed_Schema::Portal     { i, .. }     => { i.into_type_idx() }
             NP_Parsed_Schema::Any        { i, .. }     => { i.into_type_idx() }
             NP_Parsed_Schema::UTF8String { i, .. }     => { i.into_type_idx() }
             NP_Parsed_Schema::Bytes      { i, .. }     => { i.into_type_idx() }
@@ -764,6 +773,7 @@ impl NP_Parsed_Schema {
             NP_Parsed_Schema::Map        { i, .. }     => { i.into_type_idx() }
             NP_Parsed_Schema::List       { i, .. }     => { i.into_type_idx() }
             NP_Parsed_Schema::Tuple      { i, .. }     => { i.into_type_idx() }
+            NP_Parsed_Schema::Union      { i, .. }     => { i.into_type_idx() }
         }
     }
 
@@ -795,7 +805,8 @@ impl NP_Parsed_Schema {
             NP_Parsed_Schema::Map        { sortable, .. }     => { *sortable }
             NP_Parsed_Schema::List       { sortable, .. }     => { *sortable }
             NP_Parsed_Schema::Tuple      { sortable, .. }     => { *sortable }
-            NP_Parsed_Schema::Portal  { sortable, .. }     => { *sortable }
+            NP_Parsed_Schema::Portal     { sortable, .. }     => { *sortable }
+            NP_Parsed_Schema::Union      { sortable, .. }     => { *sortable }
         }
     }
 }
@@ -804,7 +815,7 @@ impl NP_Parsed_Schema {
 
 /// New NP Schema
 #[doc(hidden)]
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct NP_Schema {
     /// is this schema sortable?
     pub is_sortable: bool,
@@ -847,7 +858,8 @@ impl NP_Schema {
             NP_Parsed_Schema::Map        { .. }      => {    NP_Map::schema_to_json(parsed_schema, address) }
             NP_Parsed_Schema::List       { .. }      => {   NP_List::schema_to_json(parsed_schema, address) }
             NP_Parsed_Schema::Tuple      { .. }      => {  NP_Tuple::schema_to_json(parsed_schema, address) }
-            NP_Parsed_Schema::Portal  { .. }      => {  NP_Portal::schema_to_json(parsed_schema, address) }
+            NP_Parsed_Schema::Portal     { .. }      => { NP_Portal::schema_to_json(parsed_schema, address) }
+            NP_Parsed_Schema::Union      { .. }      => {  NP_Union::schema_to_json(parsed_schema, address) }
             _ => { Ok(NP_JSON::Null) }
         }
     }
@@ -928,7 +940,8 @@ impl NP_Schema {
             NP_TypeKeys::Map        => {       NP_Map::from_bytes_to_schema(cache, address, bytes) }
             NP_TypeKeys::List       => {      NP_List::from_bytes_to_schema(cache, address, bytes) }
             NP_TypeKeys::Tuple      => {     NP_Tuple::from_bytes_to_schema(cache, address, bytes) }
-            NP_TypeKeys::Portal     => {     NP_Portal::from_bytes_to_schema(cache, address, bytes) }
+            NP_TypeKeys::Portal     => {    NP_Portal::from_bytes_to_schema(cache, address, bytes) }
+            NP_TypeKeys::Union      => {     NP_Union::from_bytes_to_schema(cache, address, bytes) }
         }
     }
 
@@ -986,7 +999,8 @@ impl NP_Schema {
                     "list"     => {   NP_List::from_json_to_schema(schema, &json_schema) },
                     "map"      => {    NP_Map::from_json_to_schema(schema, &json_schema) },
                     "tuple"    => {  NP_Tuple::from_json_to_schema(schema, &json_schema) },
-                    "portal"   => {  NP_Portal::from_json_to_schema(schema, &json_schema) },
+                    "portal"   => { NP_Portal::from_json_to_schema(schema, &json_schema) },
+                    "union"    => {  NP_Union::from_json_to_schema(schema, &json_schema) },
                     _ => {
                         let mut err_msg = String::from("Can't find a type that matches this schema! ");
                         err_msg.push_str(json_schema.stringify().as_str());